Abstract Curve25519
