+++ src/brs/crypto/Crypto.java
@@ -1,5 +1,7 @@
 package brs.crypto;
 
+import brs.crypto.ec.Curve25519;
+import brs.crypto.ec.Curve25519Impl;
 import brs.crypto.hash.Shabal256;
 import brs.util.Convert;
 import org.bouncycastle.crypto.CipherParameters;
@@ -23,6 +25,8 @@ public final class Crypto {
 
   private static final ThreadLocal<SecureRandom> secureRandom = ThreadLocal.withInitial(SecureRandom::new);
 
+  private static final Curve25519 curve25519 = new Curve25519Impl();
+
   private Crypto() {
   } //never
 
@@ -30,7 +34,7 @@ public final class Crypto {
     try {
       return MessageDigest.getInstance(algorithm);
     } catch (NoSuchAlgorithmException e) {
-      logger.info("Missing message shabalDigest algorithm: " + algorithm);
+      logger.info("Missing message digest algorithm: " + algorithm);
       throw new RuntimeException(e.getMessage(), e);
     }
   }
@@ -44,74 +48,21 @@ public final class Crypto {
   }
 
   public static byte[] getPublicKey(String secretPhrase) {
-    byte[] publicKey = new byte[32];
-    Curve25519.keygen(publicKey, null, Crypto.sha256().digest(Convert.toBytes(secretPhrase)));
-
-    return publicKey;
+    return curve25519.getPublicKey(Crypto.sha256().digest(Convert.toBytes(secretPhrase)));
   }
 
   public static byte[] getPrivateKey(String secretPhrase) {
     byte[] s = Crypto.sha256().digest(Convert.toBytes(secretPhrase));
-    Curve25519.clamp(s);
+    curve25519.clampPrivateKey(s);
     return s;
   }
 
-  public static void curve(byte[] Z, byte[] k, byte[] P) {
-    Curve25519.curve(Z, k, P);
-  }
-
   public static byte[] sign(byte[] message, String secretPhrase) {
-
-    byte[] P = new byte[32];
-    byte[] s = new byte[32];
-    MessageDigest digest = Crypto.sha256();
-    Curve25519.keygen(P, s, digest.digest(Convert.toBytes(secretPhrase)));
-    byte[] m = digest.digest(message);
-
-    digest.update(m);
-    byte[] x = digest.digest(s);
-
-    byte[] Y = new byte[32];
-    Curve25519.keygen(Y, null, x);
-
-    digest.update(m);
-    byte[] h = digest.digest(Y);
-
-    byte[] v = new byte[32];
-    Curve25519.sign(v, h, x, s);
-
-    byte[] signature = new byte[64];
-    System.arraycopy(v, 0, signature, 0, 32);
-    System.arraycopy(h, 0, signature, 32, 32);
-
-    return signature;
+      return curve25519.sign(message, getPrivateKey(secretPhrase));
   }
 
   public static boolean verify(byte[] signature, byte[] message, byte[] publicKey, boolean enforceCanonical) {
-
-    if (enforceCanonical && !Curve25519.isCanonicalSignature(signature)) {
-      logger.debug("Rejecting non-canonical signature");
-      return false;
-    }
-
-    if (enforceCanonical && !Curve25519.isCanonicalPublicKey(publicKey)) {
-      logger.debug("Rejecting non-canonical public key");
-      return false;
-    }
-
-    byte[] Y = new byte[32];
-    byte[] v = new byte[32];
-    System.arraycopy(signature, 0, v, 0, 32);
-    byte[] h = new byte[32];
-    System.arraycopy(signature, 32, h, 0, 32);
-    Curve25519.verify(Y, v, h, publicKey);
-
-    MessageDigest digest = Crypto.sha256();
-    byte[] m = digest.digest(message);
-    digest.update(m);
-    byte[] h2 = digest.digest(Y);
-
-    return Arrays.equals(h, h2);
+      return curve25519.verify(message, signature, publicKey, enforceCanonical);
   }
 
   public static byte[] aesEncrypt(byte[] plaintext, byte[] myPrivateKey, byte[] theirPublicKey) {
@@ -120,8 +71,7 @@ public final class Crypto {
 
   public static byte[] aesEncrypt(byte[] plaintext, byte[] myPrivateKey, byte[] theirPublicKey, byte[] nonce) {
     try {
-      byte[] dhSharedSecret = new byte[32];
-      Curve25519.curve(dhSharedSecret, myPrivateKey, theirPublicKey);
+      byte[] dhSharedSecret = curve25519.getSharedSecret(myPrivateKey, theirPublicKey);
       for (int i = 0; i < 32; i++) {
         dhSharedSecret[i] ^= nonce[i];
       }
@@ -154,8 +104,7 @@ public final class Crypto {
       }
       byte[] iv = Arrays.copyOfRange(ivCiphertext, 0, 16);
       byte[] ciphertext = Arrays.copyOfRange(ivCiphertext, 16, ivCiphertext.length);
-      byte[] dhSharedSecret = new byte[32];
-      Curve25519.curve(dhSharedSecret, myPrivateKey, theirPublicKey);
+        byte[] dhSharedSecret = curve25519.getSharedSecret(myPrivateKey, theirPublicKey);
       for (int i = 0; i < 32; i++) {
         dhSharedSecret[i] ^= nonce[i];
       }
@@ -176,9 +125,7 @@ public final class Crypto {
 
   public static byte[] getSharedSecret(byte[] myPrivateKey, byte[] theirPublicKey) {
     try {
-      byte[] sharedSecret = new byte[32];
-      Curve25519.curve(sharedSecret, myPrivateKey, theirPublicKey);
-      return sharedSecret;
+      return curve25519.getSharedSecret(myPrivateKey, theirPublicKey);
     } catch (RuntimeException e) {
       logger.info("Error getting shared secret", e);
       throw e;
